+++ pyClarion/components/__init__.py
@@ -7,9 +7,10 @@ from .stores import BLATracker, Store, GoalStore
 from .wm import Flags, Slots 
 from .filters import Gates, DimFilter
 from .networks import NAM
+from .ms import Drives
 
 
 __all__ = ["Repeat", "Receptors", "Actions", "CAM", "Shift", "BoltzmannSampler", 
     "ActionSampler", "BottomUp", "TopDown", "AssociativeRules", "ActionRules", 
     "BLATracker", "Store", "GoalStore", "Flags", "Slots", "Gates", "DimFilter", 
-    "NAM"]
+    "NAM", "Drives"]